attendence-backend/requirements/base.txt
amqp==2.1.4
appdirs==1.4.0
appnope==0.1.0
billiard==3.5.0.2
celery==4.0.2
cffi==1.9.1
coreapi==2.2.4
coreschema==0.0.4
cryptography==1.7.2
decorator==4.0.11
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.10.5
django-braces==1.11.0
django-choices==1.5.0
django-cors-headers==2.0.2
django-filter==1.0.1
django-grappelli==2.9.1
django-import-export==0.5.1
django-money==0.10.2
django-rest-swagger==2.1.1
djangorestframework==3.5.4
djangorestframework-bulk==0.2.1
djangorestframework-gis==0.11
future==0.16.0
idna==2.2
ipython==5.2.2
ipython-genutils==0.1.0
itypes==1.1.0
Jinja2==2.9.5
kombu==4.0.2
Markdown==2.6.8
MarkupSafe==0.23
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
psycopg2==2.7
ptyprocess==0.5.1
py-moneyed==0.6.0
pyasn1==0.2.2
pycparser==2.17
pycrypto==2.6.1
pycryptodomex==3.4.3
Pygments==2.2.0
pyjwkest==1.3.2
PyJWT==1.4.2
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.2
tablib==0.11.4
traitlets==4.3.1
uritemplate==3.0.0
urllib3==1.20
vine==1.1.3
wcwidth==0.1.7
